Move keyvalue to its own file
