+++ packages/doenetml-worker-rust/lib-doenetml-core/src/utils.rs
@@ -1,4 +1,7 @@
 pub mod logging;
 pub mod parse_json;
+pub mod keyvalue;
 
+#[allow(unused)]
 pub use logging::*;
+pub use keyvalue::*;
